+++ src/gc/driver.c
@@ -443,6 +443,7 @@ void gc_trace_children(mino_state_t *S, gc_hdr_t *h)
                 gc_mark_child_push(S, bc);
                 gc_mark_child_push(S, bc->code);
                 gc_mark_child_push(S, bc->consts);
+                gc_mark_child_push(S, bc->clauses);
             }
             break;
         case MINO_ATOM: